+++ R/RunDiagnostics.R
@@ -38,18 +38,6 @@
 #' @template CohortSetSpecs
 #'
 #' @template CohortSetReference
-#' @param phenotypeDescriptionFile    (Optional) The location of the phenotype description file within the package.
-#'                                    The file must be .csv file and have the following columns that may be read into following
-#'                                    data types: phenotypeId (double), phenotypeName (character),
-#'                                    referentConceptId (double), clinicalDescription (character),
-#'                                    literatureReview (character), phenotypeNotes (character). Note: the field
-#'                                    names are in snake_case. Also, character fields should not have 'NA' in it.
-#'                                    'NA's are commonly added by R when R functions exports data from dataframe
-#'                                    into CSV. Instead please use '' (empty string) to represent absence of data.
-#'                                    The literature_review field is expected to be a html link to page that contains
-#'                                    resources related to literature review for the phenotype, and will be used
-#'                                    to create link-out object. This csv file is expected to be encoded in either
-#'                                    ASCII or UTF-8, if not, an error message will be displayed and process stopped.
 #' @param inclusionStatisticsFolder   The folder where the inclusion rule statistics are stored. Can be
 #'                                    left NULL if \code{runInclusionStatistics = FALSE}.
 #' @param exportFolder                The folder where the output will be exported to. If this folder
@@ -90,7 +78,6 @@ runCohortDiagnostics <- function(packageName = NULL,
                                  cohortToCreateFile = "settings/CohortsToCreate.csv",
                                  baseUrl = NULL,
                                  cohortSetReference = NULL,
-                                 phenotypeDescriptionFile = NULL,
                                  connectionDetails = NULL,
                                  connection = NULL,
                                  cdmDatabaseSchema,
@@ -223,16 +210,6 @@ runCohortDiagnostics <- function(packageName = NULL,
     cohortSetReference = cohortSetReference,
     cohortIds = cohortIds
   )
-  
-  if (!is.null(phenotypeDescriptionFile)) {
-    loadAndExportPhenotypeDescription(
-      packageName = packageName,
-      phenotypeDescriptionFile = phenotypeDescriptionFile,
-      exportFolder = exportFolder,
-      cohorts = cohorts,
-      errorMessage = errorMessage
-    )
-  }
   
   if (nrow(cohorts) == 0) {
     stop("No cohorts specified")